Ajouter entrer.c et affichage.c
