+++ src/entrer.h
@@ -1,0 +1,6 @@
+#ifndef ENTRER_H
+#define ENTRER_H
+
+int nouvelle_touche();
+
+#endif
