+++ src/main.c
@@ -11,6 +11,7 @@
 #include "editeur.h"
 #include "pacman.h"
 #include "fantome.h"
+#include "entrer.h"
 
 /******************************************************************************/
 /* MAIN                                                                       */
@@ -133,54 +134,6 @@ void dessiner_jeu(Partie *p) {
 }
 
 // Fonctions d'aide pour vérifier si une manette est branchée et qu'une touche est appuyée
-static int touche_manete(int bouton) {
-    if (manette == NULL) return 0;
-    return SDL_JoystickGetButton(manette, bouton);
-}
-
-static Uint8 croix_manette() {
-    if (manette == NULL) return 0;
-    return SDL_JoystickGetHat(manette, 0);
-}
-
-int nouvelle_touche() {
-    static int derniere_touche = 0;
-    static Uint8 derniere_croix  = 0;
-    static int dernier_bouton = 0;
-
-    int touches[] = {SDLK_DOWN, SDLK_UP, SDLK_LEFT, SDLK_RIGHT, SDLK_RETURN, SDLK_q};
-    int croix[] = {SDL_HAT_DOWN, SDL_HAT_UP, SDL_HAT_LEFT, SDL_HAT_RIGHT};
-    int boutons[] = {0,1};
-
-    int touche = attendre_touche_duree(10);
-    Uint8 croix_presse = croix_manette();
-
-    // Détecte le pression des touches entrer et q ou 0 et 1 sur manette tout en évidant la répétition
-    int bouton = 0;
-    for (int i = 0; i < 2; i++) {
-        if (touches[4 + i] != dernier_bouton && (touche == touches[4 + i] || touche_manete(boutons[i]))) {
-            bouton = touches[4 + i];
-            break;
-        }
-    }
-
-    // Vérifie si la touche détecté était déjà appuyé lors du dernier appel, évite
-    // la répétition de la touche (ou de la croix)
-    if (touche == derniere_touche && croix_presse == derniere_croix) return 0;
-
-    derniere_touche = touche;
-    derniere_croix = croix_presse;
-
-    // Vérifie les directions
-    for (int i = 0; i < 4; i++) {
-        if (touche == touches[i] || croix_presse & croix[i]) {
-            return touches[i];
-        }
-    }
-    // Aucune touche pressée
-    return bouton;
-}
-
 void actualiser_accueil(Partie *p, Timer *t) {
     int touche = nouvelle_touche();
     if (touche == SDLK_DOWN) {
@@ -226,6 +179,7 @@ void activer_bouton(Partie *p, Timer *t) {
             break;
         }
         lancer_editeur(chemin);
+        free(chemin);
         break;
     }
     case 3: exit(0); break;
