+++ src/main.h
@@ -29,6 +29,5 @@ char* selectionner_niveau();
 int charger_niveau(Partie *p, char *chemin);
 void charger_accueil();
 void afficher_liste_niveaux(char **liste, int n, int curseur);
-int nouvelle_touche();
 
 #endif
