+++ src/partie.h
@@ -38,8 +38,5 @@ void actualiser_partie(Partie *p, Timer *timer);
 void dessiner_grille(Partie *p, int dans_editeur);
 void dessiner_texte(Partie *p);
 void dessiner_partie(Partie *p);
-void terminer_partie(Partie *p);
-char *entrer_nom();
-void afficher_nom(char *nom, int index);
 
 #endif
